move: Move package Email Domain & JWT
