+++ auth-server/src/main/java/com/huyeon/authserver/email/EmailService.java
@@ -1,4 +1,4 @@
-package com.huyeon.authserver.auth.service;
+package com.huyeon.authserver.email;
 
 public interface EmailService {
     void send(String to);
