+++ auth-server/src/main/java/com/huyeon/authserver/utils/jwt/JwtAccessDeniedHandler.java
@@ -1,4 +1,4 @@
-package com.huyeon.authserver.jwt;
+package com.huyeon.authserver.utils.jwt;
 
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.web.access.AccessDeniedHandler;
