+++ auth-server/src/main/java/com/huyeon/authserver/utils/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.huyeon.authserver.jwt;
+package com.huyeon.authserver.utils.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
