+++ auth-server/src/test/java/com/huyeon/authserver/auth/service/MailAuthTest.java
@@ -1,5 +1,6 @@
 package com.huyeon.authserver.auth.service;
 
+import com.huyeon.authserver.email.MailAuth;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
