+++ auth-server/src/test/java/com/huyeon/authserver/auth/service/MailSendTest.java
@@ -1,6 +1,7 @@
 package com.huyeon.authserver.auth.service;
 
 import com.huyeon.authserver.auth.dto.EmailDTO;
+import com.huyeon.authserver.email.MailSend;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
